Remove META.yml. Don't include MYMETA.* under version control.
Add a MANIFEST.SKIP.

For reasons not entirely clear, 'perl Makefile.PL' was generating a Makefile
which was not set up to write the 'resources' -- including the repository
location -- to the MYMETA.* files.  I compared this distro's Makefile.PL
with one known to generate good results and performed surgery accordingly.

Add a package Asterisk::Perl and list it as value of 'NAME' in Makefile.PL to
silence one 'make' warning and (hopefully) get PAUSE indexer working properly.